Move utils tests and base tests to own files and improved both. Better names for constant variables.
